+++ room.cpp
@@ -52,73 +52,6 @@ void room_t::user_t::update_last(const uuid new_sync_id) {
   last_time = chrono::steady_clock::now();
 }
 
-// room sync event
-
-room_t::sync_record_t::event_t::event_t(const uuid id, const uuid from, string type, json data)
-  : id(id), from(from), type(move(type)), data(move(data)) {}
-
-void to_json(json &j, const room_t::sync_record_t::event_t &event) {
-  j = {
-    { "id", to_string(event.id) },
-    { "from", to_string(event.from) },
-    { "type", event.type },
-    { "event", event.data }
-  };
-}
-
-// room sync
-
-room_t::sync_record_t::sync_record_t() : id(gen_id()) {}
-
-void room_t::sync_record_t::add_events(
-  const uuid from, const vector<shared_ptr<event_t>> &new_reports, const vector<shared_ptr<event_t>> &new_actions
-) {
-  lock_guard lock(record_mutex);
-  if (users_phase[from] > phase_t::CREATED) throw bad_request_error("Record already synced.");
-  for (const shared_ptr<event_t> &report: new_reports) {
-    if (report->from != from) throw bad_request_error("Invalid report from.");
-    reports[report->id] = report;
-  }
-  for (const shared_ptr<event_t> &action: new_actions) {
-    if (action->from != from) throw bad_request_error("Invalid action from.");
-    actions[action->id] = action;
-  }
-  users_phase[from] = phase_t::WAITING;
-}
-
-vector<shared_ptr<room_t::sync_record_t::event_t>> room_t::sync_record_t::get_reports() const {
-  shared_lock lock(record_mutex);
-  const auto view = reports | views::transform(
-                      [](const pair<uuid, shared_ptr<event_t>> &report) { return report.second; }
-                    );
-  return move(vector(view.begin(), view.end()));
-}
-
-vector<shared_ptr<room_t::sync_record_t::event_t>> room_t::sync_record_t::get_actions() const {
-  shared_lock lock(record_mutex);
-  const auto view = actions | views::transform(
-                      [](const pair<uuid, shared_ptr<event_t>> &action) { return action.second; }
-                    );
-  return move(vector(view.begin(), view.end()));
-}
-
-room_t::sync_record_t::phase_t room_t::sync_record_t::get_phase(const uuid user_id) {
-  lock_guard lock(record_mutex);
-  return users_phase[user_id];
-}
-
-bool room_t::sync_record_t::advance_phase(const uuid user_id, const phase_t new_phase) {
-  lock_guard lock(record_mutex);
-  if (new_phase <= users_phase[user_id]) return false;
-  users_phase[user_id] = new_phase;
-  return true;
-}
-
-room_t::sync_record_t::phase_t room_t::sync_record_t::get_max_phase() const {
-  shared_lock lock(record_mutex);
-  return ranges::max(users_phase | views::values);
-}
-
 // room
 
 room_t::room_t(
@@ -235,7 +168,7 @@ void room_t::update_info(const json &new_info) {
   info = new_info;
 }
 
-vector<shared_ptr<room_t::sync_record_t>> room_t::sync(
+vector<shared_ptr<sync_record_t>> room_t::sync(
   const uuid user_id, const vector<shared_ptr<sync_record_t::event_t>> &reports,
   const vector<shared_ptr<sync_record_t::event_t>> &actions, const chrono::milliseconds wait_timeout,
   const chrono::milliseconds sync_timeout
@@ -304,82 +237,4 @@ size_t room_t::clean_sync_records() {
       );
     }
   );
-}
-
-// room list
-
-room_list_t::room_list_t(
-  const size_t limit, const chrono::minutes lobby_lifetime, const chrono::minutes game_lifetime, logger log_error,
-  logger log_info
-)
-  : log_error(move(log_error)), log_info(move(log_info)), lobby_lifetime(lobby_lifetime), game_lifetime(game_lifetime),
-    limit(limit) {}
-
-shared_ptr<room_t> room_list_t::create(const string &version, const room_t::user_t &owner, const size_t size) {
-  lock_guard lock(rooms_mutex);
-  if (rooms.size() >= limit) throw forbidden_error(format("Room limit reached. Max room count is {}.", limit));
-  const auto room = make_shared<room_t>(version, owner, size, lobby_lifetime, game_lifetime, log_error, log_info);
-  rooms[room->id] = room;
-  log_info(
-    format(
-      "Room created: {} (version={}, owner_id={}, size={})",
-      to_string(room->id),
-      version,
-      to_string(owner.id),
-      size
-    )
-  );
-  return room;
-}
-
-shared_ptr<room_t> room_list_t::get(const uuid id) const {
-  shared_lock lock(rooms_mutex);
-  try {
-    return rooms.at(id);
-  } catch (const out_of_range &) {
-    throw not_found_error("Room not found.");
-  }
-}
-
-bool room_list_t::exists(const uuid id) const {
-  shared_lock lock(rooms_mutex);
-  return rooms.contains(id);
-}
-
-bool room_list_t::remove(const uuid id) {
-  lock_guard lock(rooms_mutex);
-  if (rooms.erase(id) > 0) {
-    log_info(format("Room removed: {}", to_string(id)));
-    return true;
-  }
-  return false;
-}
-
-size_t room_list_t::count() const {
-  shared_lock lock(rooms_mutex);
-  return rooms.size();
-}
-
-vector<shared_ptr<room_t>> room_list_t::get_all() const {
-  shared_lock lock(rooms_mutex);
-  const auto view = rooms | views::transform([](const pair<uuid, shared_ptr<room_t>> &room) { return room.second; });
-  return move(vector(view.begin(), view.end()));
-}
-
-size_t room_list_t::get_limit() const {
-  shared_lock lock(rooms_mutex);
-  return limit;
-}
-
-void room_list_t::set_limit(const size_t new_limit) {
-  lock_guard lock(rooms_mutex);
-  limit = new_limit;
-}
-
-void room_list_t::clean(const chrono::milliseconds user_timeout) {
-  for (const shared_ptr<room_t> &room: get_all()) {
-    if (!room->is_available()) remove(room->id);
-    room->kick_expired(user_timeout);
-    room->clean_sync_records();
-  }
 }
