+++ session.hpp
@@ -12,7 +12,7 @@ public:
   const boost::uuids::uuid room_id;
   const boost::uuids::uuid user_id;
 
-  explicit session_t(boost::uuids::uuid room_id, boost::uuids::uuid user_id);
+  [[nodiscard]] explicit session_t(boost::uuids::uuid room_id, boost::uuids::uuid user_id);
 
 protected:
   static boost::uuids::uuid gen_id();
@@ -24,19 +24,21 @@ public:
 
   const logger log_error, log_info;
 
-  explicit session_list_t(logger log_error = [](const std::string &) {}, logger log_info = [](const std::string &) {});
+  [[nodiscard]] explicit session_list_t(
+    logger log_error = [](const std::string &) {}, logger log_info = [](const std::string &) {}
+  );
 
-  session_t create(const boost::uuids::uuid &room_id, const boost::uuids::uuid &user_id);
+  [[nodiscard]] session_t create(const boost::uuids::uuid &room_id, const boost::uuids::uuid &user_id);
 
-  session_t get(boost::uuids::uuid id) const;
+  [[nodiscard]] session_t get(boost::uuids::uuid id) const;
 
-  bool exists(boost::uuids::uuid id) const;
+  [[nodiscard]] bool exists(boost::uuids::uuid id) const;
 
   bool remove(boost::uuids::uuid id);
 
   size_t clean(const std::function<bool(const session_t &)> &is_expired);
 
 protected:
-  std::map<boost::uuids::uuid, session_t> sessions;
   mutable std::shared_mutex sessions_mutex;
+  std::map<boost::uuids::uuid, session_t> sessions;
 };
